Armo una ventanita que muestra las opciones de unirse a partida y crear juego, con los respectivos combos de partida y mapa. Agrego más lógica referida al manejo de requests en el server, pero está todavía a medio hacer. Voy a seguir trabajando sobre eso, para que se pueda obtener el listado de mapas y partidas desde el server, una vez que tenga lista la creación de partidas.

+++ src/sdl/Tile.cpp
@@ -55,7 +55,7 @@ void Tile::renderSprite(SDL_Rect &camera,
 void Tile::handleEvent(SDL_Event &e, std::string &desc) {
     //If mouse event happened
     if (e.type == SDL_MOUSEBUTTONDOWN) {
-        type = static_cast<int>((type == static_cast<int>(LTileSpriteMouseEvent::BUTTON_SPRITE_MOUSE_DOWN))
+        type = ((type == LTileSpriteMouseEvent::BUTTON_SPRITE_MOUSE_DOWN)
                                     ? LTileSpriteMouseEvent::BUTTON_SPRITE_DEFAULT
                                     : LTileSpriteMouseEvent::BUTTON_SPRITE_MOUSE_DOWN);
         desc.append("cambio tile...");
